dataset version 3

+++ zero_shot_segmentation/test_iou_of_zero_shot_with_prompts.py
@@ -543,7 +543,7 @@ if __name__ == "__main__":
     elif not args.output_dir:
         print("Please specify output dir.")
     else:
-        if args.remove_output_dir:
+        if args.remove_output_dir and os.path.exists(args.output_dir):
             shutil.rmtree(args.output_dir)
         main(args)
